including gitreports form
Git Reports is a free service that lets you set up a stable URL for anonymous users to submit bugs and other Issues to your GitHub repositories, without exposing your code, requiring users to log in with GitHub, or using a messy two-repository solution. It works with public and private repositories, and it works with personal and organization repositories. It's super easy to set up: log in, allow access, and activate your repository from the list! Check out the tutorial for more details. 

Allows write access for https://github.com/schneidmaster/